Moves the LocalizationDictionary typealias to revenuecatui.
